Add evaluation of CAEX fragments to the AASXRestServer
This implements an algorithm specific to CAEX-based fragments as they are defined in the MTP submodel template. This algorithm is able to load a CAEX document and return a certain element, e.g. based on the following route: 
.../aas/TypeAAS/submodels/ModuleTypePackage/elements/MTPFile/fragment/CAEX@ModuleTypePackage/BPXX_Freelance

Additionally, routes can be ended with specific modifiers to return the ID of an element (@ID), the value of an attribute (@Value) or the list of child elements, interfaces or attributes of an element (@Elements, @Interfaces, @Attributes).

This adds dependencies to the Aml.Engine in order to be able to parse a CAEX file.